Candidate fix for open file descriptor issue
When using FTP PASV mode formats, many ephemeral ports are used and
never released.  This, in turn, causes use of all available file descriptors
and eventual crash when sending large volumes of data.

- Fixes above by explicitly instructing the channel associated with the
stopped model to stop listening on the given port when execution of the
PIOA is halted.

Fixes #55