Candidate fix for non-conforming message lengths.
- Calculates a minimum message length based on required record layer
header sizes, ciphertext expansion, and minimum data storage.
- If sampled message length is less than minimum, then use simple
non-info carrying ranking.  Otherwise, use FTE with fixed slice
under the assumption that the format ends in '.+'
- Also changes function signature of __init__ by making minimum length
optional and providing the option of specifying the distribution to
be sampled from as a dictionary parameter.

This is still not tested against message loss issues, but will be
before the PR is made.

Fixes #34